archive/pl-implementation/requirements.txt
absl-py==0.11.0
antlr4-python3-runtime==4.8
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
einops==0.3.0
entrypoints==0.3
fsspec==0.8.5
future==0.18.2
fvcore==0.1.2.post20210128
gitdb==4.0.5
gitpython==3.1.12
google-auth-oauthlib==0.4.2
google-auth==1.24.0
grpcio==1.35.0
hydra-core==1.0.5
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
importlib-resources==5.1.0
iopath==0.1.3
ipykernel==5.4.3
ipython-genutils==0.2.0
ipython==7.19.0
ipywidgets==7.6.3
jedi==0.18.0
jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
kornia==0.4.1
lmdb==1.0.0
markdown==3.3.3
markupsafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numpy==1.20.0
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.5.1.48
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.1.0
plotly==4.14.3
portalocker==2.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
ptyprocess==0.7.0
py==1.10.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==1.0.8
pytz==2021.1
pywavelets==1.1.1
pywinpty==0.5.7
pyyaml==5.4.1
pyzmq==22.0.2
qtconsole==5.0.2
qtpy==1.9.0
requests-oauthlib==1.3.0
requests==2.25.1
retrying==1.3.3
rsa==4.7
scikit-image==0.18.1
scipy==1.6.0
send2trash==1.5.0
six==1.15.0
smmap==3.0.5
tabulate==0.8.7
tensorboard-plugin-wit==1.8.0
tensorboard==2.4.1
termcolor==1.1.0
terminado==0.9.2
test-tube==0.7.5
testpath==0.4.4
tifffile==2021.1.14
torch==1.7.1
torchvision==0.8.2
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==1.0.1
widgetsnbextension==3.5.1
yacs==0.1.8
zipp==3.4.0
